Moved gevent-dht to gevent_dht, added api.py and importing works.
Specifically you can now do:
from gevent_dht import distributedHashTable
and then go
a = distributedHashTable(bootstrap_node)
a['blah'] = "value"

Note that there is no duplication of keys so it may get lost if a node dies
while holding it.

This is not highly persistent. It does not do any housecleaning however so if
the node stays up it will not vanish.

This should be added soon.